Use file size to determine amount of continuous data in the file
Before this commit, loadContinuous would pre-allocate a large numpy array to
read data into, and then truncate the array once the file was finished. However,
if you opened a file with more samples than MAX_NUMBER_OF_CONTINUOUS_SAMPLES, it
would crash without explanation.

As it is possible to calculate the number of samples inside of a file from its
file size, I have changed loadContinuous to do this calculation and then
pre-allocate only the number of samples that is needed. This prevents the
function from erroring when opening large files, and is also more efficient.

This should solve #44.