Fix python2 style dict key access in OpenEphys.py
The pack function of OpenEphys.py uses `data.keys()[0]` to access a random key in
the dict data. This is valid in python 2, but in python 3 keys returns a
dict_keys object which does not support indexing. In order to get a random key
from the dict in python 3, I have replaced that code with `next(iter(data))`.